+++ graph_programs/ball5.c
@@ -15,20 +15,20 @@ Program to bounce animated balls around screen
 
 typedef struct balrec { /* ball data record */
 
-    int      x, y;   /* current position */
-    int      lx, ly; /* last position */
-    int      xd, yd; /* deltas */
+    long     x, y;   /* current position */
+    long     lx, ly; /* last position */
+    long     xd, yd; /* deltas */
     ami_color c;      /* ami_color */
    
 } balrec;
    
-int    cd;              /* current display flip select */
+long   cd;              /* current display flip select */
 balrec baltbl[MAXBALL]; /* ball data table */
 int    i;               /* index for table */
-int    nx, ny;          /* temp coordinates holders */
+long   nx, ny;          /* temp coordinates holders */
 int    rc;              /* repetition counter */
-int    ballsize;        /* size of ball onscreen */
-int    halfball;        /* half size of ball */
+long   ballsize;        /* size of ball onscreen */
+long   halfball;        /* half size of ball */
 
 int chkbrk(void)
 
@@ -46,7 +46,7 @@ int chkbrk(void)
 
 }
 
-void drawball(ami_color c, int x, int y)
+void drawball(ami_color c, long x, long y)
 
 {
 
@@ -57,11 +57,11 @@ void drawball(ami_color c, int x, int y)
    
 /* Find random number between 0 and N. */
 
-static int randn(int limit)
+static long randn(long limit)
 
 {
 
-    return ((long)limit+1)*rand()/RAND_MAX;
+    return (limit+1)*rand()/RAND_MAX;
 
 }
