+++ graph_programs/line1.c
@@ -14,7 +14,7 @@ Program to draw random lines on screen
 #define ACCEL       5
 #define COLORCHANGE 300
 
-static int x1, y1, xd1, yd1, i, x2, y2, xd2, yd2, lx1, ly1, lx2, ly2;
+static long x1, y1, xd1, yd1, i, x2, y2, xd2, yd2, lx1, ly1, lx2, ly2;
 static int cc; /* color counter */
 static ami_color clr;
 
@@ -39,11 +39,11 @@ static int chkbrk(void)
 
 /* Find random number between 0 and N. */
 
-static int randn(int limit)
+static long randn(long limit)
 
 {
 
-    return (long)limit*rand()/RAND_MAX;
+    return limit*rand()/RAND_MAX;
 
 }
