+++ include/services.h
@@ -30,7 +30,7 @@ extern "C" {
 #define INCSET(s, b) (!!(s[b>>3] & BIT(b%8))) /* test inclusion */
 #define ADDCSET(s, b) (s[b>>3] |= BIT(b%8)) /* add set member */
 #define SUBCSET(s, b) (s[b>>3] &= BITMSK(b%8)) /* remove set member */
-#define CLRCSET(s) { int i; for (i = 0; i < SETLEN; i++) s[i] = 0; } /* clear set */
+#define CLRCSET(s) { long i; for (i = 0; i < SETLEN; i++) s[i] = 0; } /* clear set */
 
 /*
  * Set manipulation operators for integer sets.
@@ -105,32 +105,32 @@ typedef unsigned char ami_chrset[CSETLEN];
  * Functions exposed in the services module
  */
 extern void ami_list(char* f, ami_filrec **lp);
-extern void ami_times(char* s, int sl, int t);
-extern void ami_dates(char* s, int sl, int t);
-extern void ami_writetime(FILE *f, int t);
-extern void ami_writedate(FILE *f, int t);
+extern void ami_times(char* s, long sl, long t);
+extern void ami_dates(char* s, long sl, long t);
+extern void ami_writetime(FILE *f, long t);
+extern void ami_writedate(FILE *f, long t);
 extern long ami_time(void);
 extern long ami_local(long t);
 extern long ami_clock(void);
 extern long ami_elapsed(long r);
-extern int  ami_validfile(char* s);
-extern int  ami_validpath(char* s);
-extern int  ami_wild(char* s);
-extern void ami_getenv(char* ls, char* ds, int dsl);
+extern long  ami_validfile(char* s);
+extern long  ami_validpath(char* s);
+extern long  ami_wild(char* s);
+extern void ami_getenv(char* ls, char* ds, long dsl);
 extern void ami_setenv(char* sn, char* sd);
 extern void ami_allenv(ami_envrec **el);
 extern void ami_remenv(char* sn);
 extern void ami_exec(char* cmd);
 extern void ami_exece(char* cmd, ami_envrec *el);
-extern void ami_execw(char* cmd, int *e);
-extern void ami_execew(char* cmd, ami_envrec *el, int *e);
-extern void ami_getcur(char* fn, int l);
+extern void ami_execw(char* cmd, long *e);
+extern void ami_execew(char* cmd, ami_envrec *el, long *e);
+extern void ami_getcur(char* fn, long l);
 extern void ami_setcur(char* fn);
-extern void ami_brknam(char* fn, char* p, int pl, char* n, int nl, char* e, int el);
-extern void ami_maknam(char* fn, int fnl, char* p, char* n, char* e);
-extern void ami_fulnam(char* fn, int fnl);
-extern void ami_getpgm(char* p, int pl);
-extern void ami_getusr(char* fn, int fnl);
+extern void ami_brknam(char* fn, char* p, long pl, char* n, long nl, char* e, long el);
+extern void ami_maknam(char* fn, long fnl, char* p, char* n, char* e);
+extern void ami_fulnam(char* fn, long fnl);
+extern void ami_getpgm(char* p, long pl);
+extern void ami_getusr(char* fn, long fnl);
 extern void ami_setatr(char* fn, ami_attrset a);
 extern void ami_resatr(char* fn, ami_attrset a);
 extern void ami_bakupd(char* fn);
@@ -145,33 +145,33 @@ extern void ami_rempth(char* fn);
 extern void ami_filchr(ami_chrset fc);
 extern char ami_optchr(void);
 extern char ami_pthchr(void);
-extern int  ami_latitude(void);
-extern int  ami_longitude(void);
-extern int  ami_altitude(void);
-extern int  ami_country(void);
-extern void ami_countrys(char* s, int sl, int c);
-extern int  ami_timezone(void);
-extern int  ami_daysave(void);
-extern int  ami_time24hour(void);
-extern int  ami_language(void);
-extern void ami_languages(char* s, int sl, int l);
+extern long  ami_latitude(void);
+extern long  ami_longitude(void);
+extern long  ami_altitude(void);
+extern long  ami_country(void);
+extern void ami_countrys(char* s, long sl, long c);
+extern long  ami_timezone(void);
+extern long  ami_daysave(void);
+extern long  ami_time24hour(void);
+extern long  ami_language(void);
+extern void ami_languages(char* s, long sl, long l);
 extern char ami_decimal(void);
 extern char ami_numbersep(void);
-extern int  ami_timeorder(void);
-extern int  ami_dateorder(void);
+extern long  ami_timeorder(void);
+extern long  ami_dateorder(void);
 extern char ami_datesep(void);
 extern char ami_timesep(void);
 extern char ami_currchr(void);
-extern int ami_newthread(void (*threadmain)(void));
-extern int ami_initlock(void);
-extern void ami_deinitlock(int ln);
-extern void ami_lock(int ln);
-extern void ami_unlock(int ln);
-extern int ami_initsig(void);
-extern void ami_deinitsig(int sn);
-extern void ami_sendsig(int sn);
-extern void ami_sendsigone(int sn);
-extern void ami_waitsig(int ln, int sn);
+extern long ami_newthread(void (*threadmain)(void));
+extern long ami_initlock(void);
+extern void ami_deinitlock(long ln);
+extern void ami_lock(long ln);
+extern void ami_unlock(long ln);
+extern long ami_initsig(void);
+extern void ami_deinitsig(long sn);
+extern void ami_sendsig(long sn);
+extern void ami_sendsigone(long sn);
+extern void ami_waitsig(long ln, long sn);
 
 #ifdef __cplusplus
 }
