+++ linux/system_event.c
@@ -32,7 +32,7 @@
 * an event when the signal occurs. The logical system event number is          *
 * returned.                                                                    *
 *                                                                              *
-* int system_event_addsetim(int sid, int t, int r);                            *
+* int system_event_addsetim(int sid, long t, long r);                          *
 *                                                                              *
 * Activates a timer with the given time and repeat state. An event occurs when *
 * the timer fires, and may repeat. Both accepts a logical system event number  *
@@ -287,7 +287,7 @@ returned.
 
 *******************************************************************************/
 
-int system_event_addsetim(int sid, int t, int r)
+int system_event_addsetim(int sid, long t, long r)
 
 {
